% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{curiesr-package}
\alias{curiesr}
\alias{curiesr-package}
\title{curiesr: Idiomatic conversion between URIs and compact URIs (CURIEs)}
\description{
R package for the curies library: idiomatic conversion between URIs and compact URIs (CURIEs)
}
\author{
\strong{Maintainer}: Vincent Emonet \email{vincent.emonet@gmail.com} (\href{https://orcid.org/0000-0002-1501-1082}{ORCID})

Authors:
\itemize{
  \item Charles Tapley Hoyt \email{cthoyt@gmail.com}
}

}
